Move hint initialization to a new module
